added bought sold info quiery

+++ db/schema.sql
@@ -25,7 +25,7 @@ CREATE TABLE specs (
     color VARCHAR(30) NOT NULL,
     transmission VARCHAR(30) NOT NULL,
     bought DECIMAL(10, 2) NOT NULL,
-    start_mile DECIMAL(10, 2) NOT NULL,
+    start_mile INT NOT NULL,
     sold DECIMAL(10, 2),
     end_mileage INT,
     model_id INT NOT NULL,
